refactor: migrate add product to quote request component
