+++ src/app/extensions/quoting/exports/quoting-exports.module.ts
@@ -3,8 +3,6 @@ import { NgModule } from '@angular/core';
 import { FeatureToggleModule } from 'ish-core/feature-toggle.module';
 import { LAZY_FEATURE_MODULE } from 'ish-core/utils/module-loader/module-loader.service';
 
-import { LazyProductAddToQuoteComponent } from './lazy-product-add-to-quote/lazy-product-add-to-quote.component';
-
 @NgModule({
   imports: [FeatureToggleModule],
   providers: [
@@ -14,7 +12,7 @@ import { LazyProductAddToQuoteComponent } from './lazy-product-add-to-quote/lazy
       multi: true,
     },
   ],
-  declarations: [LazyProductAddToQuoteComponent],
-  exports: [LazyProductAddToQuoteComponent],
+  declarations: [],
+  exports: [],
 })
 export class QuotingExportsModule {}
