+++ src/app/extensions/quoting2/exports/quoting2-exports.module.ts
@@ -4,6 +4,7 @@ import { FeatureToggleModule } from 'ish-core/feature-toggle.module';
 import { LAZY_FEATURE_MODULE } from 'ish-core/utils/module-loader/module-loader.service';
 
 import { LazyBasketAddToQuoteComponent } from './lazy-basket-add-to-quote/lazy-basket-add-to-quote.component';
+import { LazyProductAddToQuoteComponent } from './lazy-product-add-to-quote/lazy-product-add-to-quote.component';
 import { LazyQuoteWidgetComponent } from './lazy-quote-widget/lazy-quote-widget.component';
 
 @NgModule({
@@ -15,7 +16,7 @@ import { LazyQuoteWidgetComponent } from './lazy-quote-widget/lazy-quote-widget.
       multi: true,
     },
   ],
-  declarations: [LazyBasketAddToQuoteComponent, LazyQuoteWidgetComponent],
-  exports: [LazyBasketAddToQuoteComponent, LazyQuoteWidgetComponent],
+  declarations: [LazyBasketAddToQuoteComponent, LazyProductAddToQuoteComponent, LazyQuoteWidgetComponent],
+  exports: [LazyBasketAddToQuoteComponent, LazyProductAddToQuoteComponent, LazyQuoteWidgetComponent],
 })
 export class Quoting2ExportsModule {}
