+++ src/app/extensions/quoting2/quoting2.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 import { SharedModule } from 'ish-shared/shared.module';
 
 import { BasketAddToQuoteComponent } from './shared/basket-add-to-quote/basket-add-to-quote.component';
+import { ProductAddToQuoteComponent } from './shared/product-add-to-quote/product-add-to-quote.component';
 import { QuoteEditComponent } from './shared/quote-edit/quote-edit.component';
 import { QuoteExpirationDateComponent } from './shared/quote-expiration-date/quote-expiration-date.component';
 import { QuoteStateComponent } from './shared/quote-state/quote-state.component';
@@ -13,6 +14,7 @@ import { QuoteWidgetComponent } from './shared/quote-widget/quote-widget.compone
   imports: [SharedModule],
   declarations: [
     BasketAddToQuoteComponent,
+    ProductAddToQuoteComponent,
     QuoteEditComponent,
     QuoteExpirationDateComponent,
     QuoteStateComponent,
@@ -20,6 +22,7 @@ import { QuoteWidgetComponent } from './shared/quote-widget/quote-widget.compone
     QuoteWidgetComponent,
   ],
   exports: [
+    ProductAddToQuoteComponent,
     QuoteEditComponent,
     QuoteExpirationDateComponent,
     QuoteStateComponent,
