+++ src/app/shared/components/product/product-row/product-row.component.spec.ts
@@ -23,7 +23,7 @@ import { ProductVariationSelectComponent } from 'ish-shared/components/product/p
 import { ProductImageComponent } from 'ish-shell/header/product-image/product-image.component';
 
 import { LazyProductAddToOrderTemplateComponent } from '../../../../extensions/order-templates/exports/lazy-product-add-to-order-template/lazy-product-add-to-order-template.component';
-import { LazyProductAddToQuoteComponent } from '../../../../extensions/quoting/exports/lazy-product-add-to-quote/lazy-product-add-to-quote.component';
+import { LazyProductAddToQuoteComponent } from '../../../../extensions/quoting2/exports/lazy-product-add-to-quote/lazy-product-add-to-quote.component';
 import { IsTactonProductDirective } from '../../../../extensions/tacton/directives/is-tacton-product.directive';
 import { LazyTactonConfigureProductComponent } from '../../../../extensions/tacton/exports/lazy-tacton-configure-product/lazy-tacton-configure-product.component';
 import { LazyProductAddToWishlistComponent } from '../../../../extensions/wishlists/exports/lazy-product-add-to-wishlist/lazy-product-add-to-wishlist.component';
